Store contributors as an array and store math functions in a seperate file
