+++ CONTRIBUTORS.js
@@ -2,14 +2,13 @@
 * This file is for contributors to Slothpixel to add their Minecraft UUID.
 * Currently this will cause the /player/:id endpoint to return is_contributor = true for your profile
 */
-
-module.exports = {
-  ef962ec2df6e48a2ac9d6062c1b84652: {}, // builder_247
-  '7beb3348294f44d8b967cdf0d872aff8': {}, // ecolsson
-  '63fd5c413cd641e987d473df9c0e1e56': {}, // ThePoptartCrpr
-  '532de453ac214ddeaec9a18a24a79e27': {}, // Cham
-  f8a9e651e8544abd8fef5a02a75ac25a: {}, // DidiSkywalker
-  '27ae0f1720894bd7a11b08e741aa9024': {},
-  e2db3b87ae5c4b91a04f7d6f5ef51e27: {}, // zunora
-  '56da43a4088d4a7682b6dd431535015e': {}, // Richienb
-};
+module.exports = [
+  'ef962ec2df6e48a2ac9d6062c1b84652', // builder_247
+  '7beb3348294f44d8b967cdf0d872aff8', // ecolsson
+  '63fd5c413cd641e987d473df9c0e1e56', // ThePoptartCrpr
+  '532de453ac214ddeaec9a18a24a79e27', // Cham
+  'f8a9e651e8544abd8fef5a02a75ac25a', // DidiSkywalker
+  '27ae0f1720894bd7a11b08e741aa9024', // Hardstyles
+  'e2db3b87ae5c4b91a04f7d6f5ef51e27', // zunora
+  '56da43a4088d4a7682b6dd431535015e', // Richienb
+];
